+++ src/index.html
@@ -1,12 +1,44 @@
 <!DOCTYPE html>
-<html>
-  <head>
+<html lang="en">
+<head>
     <meta charset="UTF-8">
+    <!-- https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP -->
+    <meta http-equiv="Content-Security-Policy" content="script-src 'self'">
+    <meta http-equiv="X-Content-Security-Policy" content="script-src 'self'">
+    <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <link href="assets/style.css" rel="stylesheet"/>
     <title>Hello World!</title>
+    <style>
+        * {
+            box-sizing: border-box;
+        }
 
-  </head>
-  <body>
-    <h1>💖 Hello World!</h1>
-    <p>Welcome to your Electron application.</p>
-  </body>
+        html {
+            height: 100%;
+        }
+
+        body {
+            height: 100%;
+            background: #3A3A3A;
+            color: #EEEEEE;
+            margin: 0;
+        }
+    </style>
+</head>
+<body>
+<ani-viewer-app></ani-viewer-app>
+
+<script src="src/shared/templates.js"></script>
+<script src="src/shared/helpers.js"></script>
+<script src="src/components/base.js"></script>
+<script src="src/components/app.js"></script>
+<script src="src/components/content.js"></script>
+<script src="src/components/details.js"></script>
+<script src="src/components/favorites.js"></script>
+<script src="src/components/header.js"></script>
+<script src="src/components/history.js"></script>
+<script src="src/components/search-results.js"></script>
+<script src="src/components/side-nav.js"></script>
+</body>
 </html>
