--- src/shared/templates.js
@@ -1,2 +1,0 @@
-const css = (strings) => strings;
-const html = (strings) => strings;
