+++ src/shared/templates.ts
@@ -1,0 +1,2 @@
+const css = (strings: TemplateStringsArray) => strings;
+const html = (strings: TemplateStringsArray) => strings;
